refactor(cli): rename test directory for consistency
